Thêm các endpoint cho variant, hoàn thiện thêm về checkout flow

+++ src/Services/Orders/Ecommerce.Services.Orders.Contracts/Events/OrderCreatedEvent.cs
@@ -9,7 +9,7 @@ public record OrderCreatedEvent : IIntegrationEvent
     public Guid OrderId { get; set; }
     public DateTime CreatedAt { get; set; }
     public long CustomerId { get; set; }
-    public long TotalAmount { get; set; }
+    public decimal TotalAmount { get; set; }
     public long PaymentMethodId { get; set; }
     public string ShippingAddress { get; set; } = string.Empty;
     public List<OrderItemData> OrderItems { get; set; } = new List<OrderItemData>();
